Improvements.
Change API so all modifications to reports starts with a report id.

Add support for DM to bot - this solves issue with requests partially being seen in public channels.

Add locations versus trails and add missing trails/locations to list.

Change 'at' caching to use explicit cache so we can check if response it cached - and
not print the 'please wait' message.