Fix dynamo keys to support changing update time. Add deleting invoking message.
Change indexes around - have reportid as the single PK, and use constant/update_ts as a GSI
since changing a part of the PK of course causes a new record to be written rather than updating.

Added deleting invoking messages like 'at'. For some reason this works for my userid but not
Karen - need to look into it more.