Add web view
Using Flask-Admin - provide an admin view - which is just a table of reports.

Can sort, filter on many columns.

Create a proxy_auth layer with Flask-Login that authenticates user based on plweb credentials.

Improve plweb login - really the only way to tell is login was successful was is we get a session cookie.